+++ postgresql/config.go
@@ -26,6 +26,7 @@ const (
 	featureSchemaCreateIfNotExist
 	featureReplication
 	featureExtension
+	featurePrivileges
 )
 
 type dbRegistryEntry struct {
@@ -65,6 +66,10 @@ var (
 
 		// CREATE EXTENSION support.
 		featureExtension: semver.MustParseRange(">=9.1.0"),
+
+		// We do not support postgresql_grant and postgresql_default_privileges
+		// for Postgresql < 9.
+		featurePrivileges: semver.MustParseRange(">=9.0.0"),
 	}
 )
 
@@ -72,7 +77,6 @@ var (
 type Config struct {
 	Host              string
 	Port              int
-	Database          string
 	Username          string
 	Password          string
 	DatabaseUsername  string
@@ -90,6 +94,8 @@ type Client struct {
 	// Configuration for the client
 	config Config
 
+	databaseName string
+
 	// db is a pointer to the DB connection.  Callers are responsible for
 	// releasing their connections.
 	db *sql.DB
@@ -106,12 +112,12 @@ type Client struct {
 	catalogLock sync.RWMutex
 }
 
-// NewClient returns new client config
-func (c *Config) NewClient() (*Client, error) {
+// NewClient returns client config for the specified database.
+func (c *Config) NewClient(database string) (*Client, error) {
 	dbRegistryLock.Lock()
 	defer dbRegistryLock.Unlock()
 
-	dsn := c.connStr()
+	dsn := c.connStr(database)
 	dbEntry, found := dbRegistry[dsn]
 	if !found {
 		db, err := sql.Open("postgres", dsn)
@@ -119,8 +125,10 @@ func (c *Config) NewClient() (*Client, error) {
 			return nil, errwrap.Wrapf("Error connecting to PostgreSQL server: {{err}}", err)
 		}
 
-		// only one connection
-		db.SetMaxIdleConns(1)
+		// We don't want to retain connection
+		// So when we connect on a specific database which might be managed by terraform,
+		// we don't keep opened connection in case of the db has to be dopped in the plan.
+		db.SetMaxIdleConns(0)
 		db.SetMaxOpenConns(c.MaxConns)
 
 		version, err := fingerprintCapabilities(db)
@@ -137,9 +145,10 @@ func (c *Config) NewClient() (*Client, error) {
 	}
 
 	client := Client{
-		config:  *c,
-		db:      dbEntry.db,
-		version: dbEntry.version,
+		config:       *c,
+		databaseName: database,
+		db:           dbEntry.db,
+		version:      dbEntry.version,
 	}
 
 	return &client, nil
@@ -158,7 +167,7 @@ func (c *Config) featureSupported(name featureName) bool {
 	return fn(c.ExpectedVersion)
 }
 
-func (c *Config) connStr() string {
+func (c *Config) connStr(database string) string {
 	// NOTE: dbname must come before user otherwise dbname will be set to
 	// user.
 	var dsnFmt string
@@ -213,7 +222,7 @@ func (c *Config) connStr() string {
 		logValues := []interface{}{
 			quote(c.Host),
 			c.Port,
-			quote(c.Database),
+			quote(database),
 			quote(c.Username),
 			quote("<redacted>"),
 			quote(c.SSLMode),
@@ -232,7 +241,7 @@ func (c *Config) connStr() string {
 		connValues := []interface{}{
 			quote(c.Host),
 			c.Port,
-			quote(c.Database),
+			quote(database),
 			quote(c.Username),
 			quote(c.Password),
 			quote(c.SSLMode),