+++ postgresql/helpers.go
@@ -2,11 +2,13 @@ package postgresql
 
 import (
 	"fmt"
+	"log"
 	"strings"
 
 	"database/sql"
 
 	"github.com/hashicorp/errwrap"
+	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/lib/pq"
 )
 
@@ -90,4 +92,112 @@ func revokeRoleMembership(db *sql.DB, role, member string) error {
 		}
 	}
 	return nil
+}
+
+func sliceContainsStr(haystack []string, needle string) bool {
+	for _, s := range haystack {
+		if s == needle {
+			return true
+		}
+	}
+	return false
+}
+
+// allowedPrivileges is the list of privileges allowed per object types in Postgres.
+// see: https://www.postgresql.org/docs/current/sql-grant.html
+var allowedPrivileges = map[string][]string{
+	"table":    []string{"ALL", "SELECT", "INSERT", "UPDATE", "DELETE", "TRUNCATE", "REFERENCES", "TRIGGER"},
+	"sequence": []string{"ALL", "USAGE", "SELECT", "UPDATE"},
+}
+
+// validatePrivileges checks that privileges to apply are allowed for this object type.
+func validatePrivileges(objectType string, privileges []interface{}) error {
+	allowed, ok := allowedPrivileges[objectType]
+	if !ok {
+		return fmt.Errorf("unknown object type %s", objectType)
+	}
+
+	for _, priv := range privileges {
+		if !sliceContainsStr(allowed, priv.(string)) {
+			return fmt.Errorf("%s is not an allowed privilege for object type %s", priv, objectType)
+		}
+	}
+	return nil
+}
+
+func pgArrayToSet(arr pq.ByteaArray) *schema.Set {
+	s := make([]interface{}, len(arr))
+	for i, v := range arr {
+		s[i] = string(v)
+	}
+	return schema.NewSet(schema.HashString, s)
+}
+
+// startTransaction starts a new DB transaction on the specified database.
+// If the database is specified and different from the one configured in the provider,
+// it will create a new connection pool if needed.
+func startTransaction(client *Client, database string) (*sql.Tx, error) {
+	if database != "" && database != client.databaseName {
+		var err error
+		client, err = client.config.NewClient(database)
+		if err != nil {
+			return nil, err
+		}
+	}
+	db := client.DB()
+	txn, err := db.Begin()
+	if err != nil {
+		return nil, errwrap.Wrapf("could not start transaction: {{err}}", err)
+	}
+
+	return txn, nil
+}
+
+func dbExists(txn *sql.Tx, dbname string) (bool, error) {
+	err := txn.QueryRow("SELECT datname FROM pg_database WHERE datname=$1", dbname).Scan(&dbname)
+	switch {
+	case err == sql.ErrNoRows:
+		return false, nil
+	case err != nil:
+		return false, errwrap.Wrapf("could not check if database exists: {{err}}", err)
+	}
+
+	return true, nil
+}
+
+func roleExists(txn *sql.Tx, rolname string) (bool, error) {
+	err := txn.QueryRow("SELECT 1 FROM pg_roles WHERE rolname=$1", rolname).Scan(&rolname)
+	switch {
+	case err == sql.ErrNoRows:
+		return false, nil
+	case err != nil:
+		return false, errwrap.Wrapf("could not check if role exists: {{err}}", err)
+	}
+
+	return true, nil
+}
+
+func schemaExists(txn *sql.Tx, schemaname string) (bool, error) {
+	err := txn.QueryRow("SELECT 1 FROM pg_namespace WHERE nspname=$1", schemaname).Scan(&schemaname)
+	switch {
+	case err == sql.ErrNoRows:
+		return false, nil
+	case err != nil:
+		return false, errwrap.Wrapf("could not check if schema exists: {{err}}", err)
+	}
+
+	return true, nil
+}
+
+// deferredRollback can be used to rollback a transaction in a defer.
+// It will log an error if it fails
+func deferredRollback(txn *sql.Tx) {
+	err := txn.Rollback()
+	switch {
+	case err == sql.ErrTxDone:
+		// transaction has already been committed or rolled back
+		log.Printf("[DEBUG]: %v", err)
+	case err != nil:
+		log.Printf("[ERR] could not rollback transaction: %v", err)
+	}
 }