+++ postgresql/utils_test.go
@@ -6,7 +6,9 @@ import (
 	"os"
 	"strconv"
 	"testing"
+	"time"
 
+	"github.com/hashicorp/errwrap"
 	"github.com/hashicorp/terraform/helper/resource"
 )
 
@@ -14,9 +16,15 @@ const (
 	dbNamePrefix     = "tf_tests_db"
 	roleNamePrefix   = "tf_tests_role"
 	testRolePassword = "testpwd"
-
-	testTableDef = "CREATE TABLE test_table (val text)"
 )
+
+// Can be used in a PreCheck function to disable test based on feature.
+func testCheckCompatibleVersion(t *testing.T, feature featureName) {
+	client := testAccProvider.Meta().(*Client)
+	if !client.featureSupported(feature) {
+		t.Skip(fmt.Sprintf("Skip extension tests for Postgres %s", client.version))
+	}
+}
 
 func getTestConfig(t *testing.T) Config {
 	getEnv := func(key, fallback string) string {
@@ -37,7 +45,6 @@ func getTestConfig(t *testing.T) Config {
 		Port:     dbPort,
 		Username: getEnv("PGUSER", ""),
 		Password: getEnv("PGPASSWORD", ""),
-		Database: getEnv("PGDATABASE", "postgres"),
 		SSLMode:  getEnv("PGSSLMODE", ""),
 	}
 }
@@ -62,4 +69,101 @@ func dbExecute(t *testing.T, dsn, query string, args ...interface{}) {
 	if _, err = db.Exec(query, args...); err != nil {
 		t.Fatalf("could not execute query %s: %v", query, err)
 	}
+}
+
+func getTestDBNames(dbSuffix string) (dbName string, roleName string) {
+	dbName = fmt.Sprintf("%s_%s", dbNamePrefix, dbSuffix)
+	roleName = fmt.Sprintf("%s_%s", roleNamePrefix, dbSuffix)
+
+	return
+}
+
+// setupTestDatabase creates all needed resources before executing a terraform test
+// and provides the teardown function to delete all these resources.
+func setupTestDatabase(t *testing.T, createDB, createRole bool) (string, func()) {
+	config := getTestConfig(t)
+
+	suffix := strconv.Itoa(int(time.Now().UnixNano()))
+
+	dbName, roleName := getTestDBNames(suffix)
+
+	if createDB {
+		dbExecute(t, config.connStr("postgres"), fmt.Sprintf("CREATE DATABASE %s", dbName))
+	}
+	if createRole {
+		dbExecute(t, config.connStr("postgres"), fmt.Sprintf(
+			"CREATE ROLE %s LOGIN ENCRYPTED PASSWORD '%s'",
+			roleName, testRolePassword,
+		))
+	}
+
+	return suffix, func() {
+		dbExecute(t, config.connStr("postgres"), fmt.Sprintf("DROP DATABASE IF EXISTS %s", dbName))
+		dbExecute(t, config.connStr("postgres"), fmt.Sprintf("DROP ROLE IF EXISTS %s", roleName))
+	}
+}
+
+func createTestTables(t *testing.T, dbSuffix string, tables []string) func() {
+	config := getTestConfig(t)
+	dbName, _ := getTestDBNames(dbSuffix)
+
+	db, err := sql.Open("postgres", config.connStr(dbName))
+	if err != nil {
+		t.Fatalf("could not open connection pool for db %s: %v", dbName, err)
+	}
+	defer db.Close()
+
+	for _, table := range tables {
+		if _, err := db.Exec(fmt.Sprintf("CREATE TABLE %s (val text)", table)); err != nil {
+			t.Fatalf("could not create test table in db %s: %v", dbName, err)
+		}
+	}
+	// In this case we need to drop table after each test.
+	return func() {
+		for _, table := range tables {
+			db.Exec(fmt.Sprintf("DROP TABLE %s", table))
+		}
+	}
+}
+
+func testCheckTablesPrivileges(t *testing.T, dbSuffix string, tables []string, allowedPrivileges []string) error {
+	config := getTestConfig(t)
+	dbName, roleName := getTestDBNames(dbSuffix)
+
+	// Connect as the test role
+	config.Username = roleName
+	config.Password = testRolePassword
+
+	db, err := sql.Open("postgres", config.connStr(dbName))
+	if err != nil {
+		t.Fatalf("could not open connection pool for db %s: %v", dbName, err)
+	}
+	defer db.Close()
+
+	for _, table := range tables {
+		queries := map[string]string{
+			"SELECT": fmt.Sprintf("SELECT count(*) FROM %s", table),
+			"INSERT": fmt.Sprintf("INSERT INTO %s VALUES ('test')", table),
+			"UPDATE": fmt.Sprintf("UPDATE %s SET val = 'test'", table),
+			"DELETE": fmt.Sprintf("DELETE FROM %s", table),
+		}
+
+		for queryType, query := range queries {
+			_, err := db.Exec(query)
+
+			if err != nil && sliceContainsStr(allowedPrivileges, queryType) {
+				return errwrap.Wrapf(
+					fmt.Sprintf("could not %s on test table %s: {{err}}", queryType, table),
+					err,
+				)
+
+			} else if err == nil && !sliceContainsStr(allowedPrivileges, queryType) {
+				return errwrap.Wrapf(
+					fmt.Sprintf("%s did not failed as expected for table %s: {{err}}", queryType, table),
+					err,
+				)
+			}
+		}
+	}
+	return nil
 }